refactor[s2-3.4]: Supported all Roles  for changing password; Moved ChangePasswordModal to main directory
