+++ Front/pharmacy/src/views/pharmacyAdmin/profile/PharmacyAdminProfileInformation.vue
@@ -77,7 +77,8 @@
 </template>
 
 <script>
-import ChangePasswordModal from './ChangePasswordModal.vue';
+import ChangePasswordModal from "@/views/ChangePasswordModal.vue";
+
 
 export default {
     props: ['user'],
